Move anchor management into separate modules
