+++ src/internet_identity/src/main.rs
@@ -1,22 +1,16 @@
 use crate::assets::init_assets;
-use crate::state::RegistrationState::{DeviceRegistrationModeActive, DeviceTentativelyAdded};
-use crate::state::{ChallengeInfo, DeviceDataInternal, TentativeDeviceRegistration};
-use crate::AddTentativeDeviceResponse::{AddedTentatively, AnotherDeviceTentativelyAdded};
-use crate::VerifyTentativeDeviceResponse::{NoDeviceToVerify, WrongCode};
+use anchor_management::tentative_device_registration;
 use assets::ContentType;
 use candid::Principal;
-use ic_cdk::api::call::call;
-use ic_cdk::api::{caller, set_certified_data, time, trap};
+use ic_cdk::api::{caller, set_certified_data, trap};
 use ic_cdk_macros::{init, post_upgrade, query, update};
 use ic_certified_map::AsHashTree;
 use ic_stable_structures::DefaultMemoryImpl;
-use rand_chacha::rand_core::{RngCore, SeedableRng};
-use std::collections::HashMap;
-use std::convert::TryInto;
 use storage::{Salt, Storage};
 
 use internet_identity_interface::*;
 
+mod anchor_management;
 mod assets;
 mod delegation;
 mod hash;
@@ -28,21 +22,6 @@ const fn secs_to_nanos(secs: u64) -> u64 {
     secs * 1_000_000_000
 }
 
-#[cfg(not(feature = "dummy_captcha"))]
-use captcha::filters::Wave;
-
-// 5 mins
-const CAPTCHA_CHALLENGE_LIFETIME: u64 = secs_to_nanos(300);
-// How many captcha challenges we keep in memory (at most)
-const MAX_INFLIGHT_CHALLENGES: usize = 500;
-
-// 15 mins
-const REGISTRATION_MODE_DURATION: u64 = secs_to_nanos(900);
-// How many users can be in registration mode simultaneously
-const MAX_USERS_IN_REGISTRATION_MODE: usize = 10_000;
-// How many verification attempts are given for a tentative device
-const MAX_DEVICE_REGISTRATION_ATTEMPTS: u8 = 3;
-
 const LABEL_ASSETS: &[u8] = b"http_assets";
 const LABEL_SIG: &[u8] = b"sig";
 
@@ -51,45 +30,14 @@ async fn init_salt() {
     state::init_salt().await;
 }
 
-/// Enables device registration mode for the given user and returns the expiration timestamp (when it will be disabled again).
-/// If the device registration mode is already active it will just return the expiration timestamp again.
 #[update]
 fn enter_device_registration_mode(user_number: UserNumber) -> Timestamp {
-    let entries = state::anchor_devices(user_number);
-    trap_if_not_authenticated(entries.iter().map(|e| &e.pubkey));
-
-    state::tentative_device_registrations_mut(|registrations| {
-        prune_expired_tentative_device_registrations(registrations);
-        if registrations.len() >= MAX_USERS_IN_REGISTRATION_MODE {
-            trap("too many users in device registration mode");
-        }
-
-        match registrations.get(&user_number) {
-            Some(TentativeDeviceRegistration { expiration, .. }) => *expiration, // already enabled, just return the existing expiration
-            None => {
-                let expiration = time() + REGISTRATION_MODE_DURATION;
-                registrations.insert(
-                    user_number,
-                    TentativeDeviceRegistration {
-                        expiration,
-                        state: DeviceRegistrationModeActive,
-                    },
-                );
-                expiration
-            }
-        }
-    })
+    tentative_device_registration::enter_device_registration_mode(user_number)
 }
 
 #[update]
 fn exit_device_registration_mode(user_number: UserNumber) {
-    let entries = state::anchor_devices(user_number);
-    trap_if_not_authenticated(entries.iter().map(|e| &e.pubkey));
-
-    state::tentative_device_registrations_mut(|registrations| {
-        prune_expired_tentative_device_registrations(registrations);
-        registrations.remove(&user_number)
-    });
+    tentative_device_registration::exit_device_registration_mode(user_number)
 }
 
 #[update]
@@ -97,34 +45,7 @@ async fn add_tentative_device(
     user_number: UserNumber,
     device_data: DeviceData,
 ) -> AddTentativeDeviceResponse {
-    let verification_code = new_verification_code().await;
-    let now = time();
-
-    state::tentative_device_registrations_mut(|registrations| {
-        prune_expired_tentative_device_registrations(registrations);
-
-        match registrations.get_mut(&user_number) {
-            None => AddTentativeDeviceResponse::DeviceRegistrationModeOff,
-            Some(TentativeDeviceRegistration { expiration, .. }) if *expiration <= now => {
-                AddTentativeDeviceResponse::DeviceRegistrationModeOff
-            }
-            Some(TentativeDeviceRegistration {
-                state: DeviceTentativelyAdded { .. },
-                ..
-            }) => AnotherDeviceTentativelyAdded,
-            Some(mut registration) => {
-                registration.state = DeviceTentativelyAdded {
-                    tentative_device: device_data,
-                    failed_attempts: 0,
-                    verification_code: verification_code.clone(),
-                };
-                AddedTentatively {
-                    device_registration_timeout: registration.expiration,
-                    verification_code,
-                }
-            }
-        }
-    })
+    tentative_device_registration::add_tentative_device(user_number, device_data).await
 }
 
 #[update]
@@ -132,370 +53,33 @@ async fn verify_tentative_device(
     user_number: UserNumber,
     user_verification_code: DeviceVerificationCode,
 ) -> VerifyTentativeDeviceResponse {
-    match get_verified_device(user_number, user_verification_code) {
-        Ok(device) => {
-            add(user_number, device).await;
-            VerifyTentativeDeviceResponse::Verified
-        }
-        Err(err) => err,
-    }
-}
-
-/// Checks the device verification code for a tentative device.
-/// If valid, returns the device to be added and exits device registration mode
-/// If invalid, returns the appropriate error to send to the client and increases failed attempts. Exits device registration mode if there are no retries left.
-fn get_verified_device(
-    user_number: UserNumber,
-    user_verification_code: DeviceVerificationCode,
-) -> Result<DeviceData, VerifyTentativeDeviceResponse> {
-    let entries = state::anchor_devices(user_number);
-    trap_if_not_authenticated(entries.iter().map(|e| &e.pubkey));
-
-    state::tentative_device_registrations_mut(|registrations| {
-        prune_expired_tentative_device_registrations(registrations);
-
-        let mut tentative_registration = registrations
-            .remove(&user_number)
-            .ok_or(VerifyTentativeDeviceResponse::DeviceRegistrationModeOff)?;
-
-        match tentative_registration.state {
-            DeviceRegistrationModeActive => Err(NoDeviceToVerify),
-            DeviceTentativelyAdded {
-                failed_attempts,
-                verification_code,
-                tentative_device,
-            } => {
-                if user_verification_code == verification_code {
-                    return Ok(tentative_device);
-                }
-
-                let failed_attempts = failed_attempts + 1;
-                if failed_attempts < MAX_DEVICE_REGISTRATION_ATTEMPTS {
-                    tentative_registration.state = DeviceTentativelyAdded {
-                        failed_attempts,
-                        tentative_device,
-                        verification_code,
-                    };
-                    // reinsert because retries are allowed
-                    registrations.insert(user_number, tentative_registration);
-                }
-                return Err(WrongCode {
-                    retries_left: (MAX_DEVICE_REGISTRATION_ATTEMPTS - failed_attempts),
-                });
-            }
-        }
-    })
-}
-
-/// Return a decimal representation of a random `u32` to be used as verification code
-async fn new_verification_code() -> DeviceVerificationCode {
-    let res: Vec<u8> = match call(Principal::management_canister(), "raw_rand", ()).await {
-        Ok((res,)) => res,
-        Err((_, err)) => trap(&format!("failed to get randomness: {}", err)),
-    };
-    let rand = u32::from_be_bytes(res[..4].try_into().unwrap_or_else(|_| {
-        trap(&format!(
-            "when creating device verification code from raw_rand output, expected raw randomness to be of length 32, got {}",
-            res.len()
-        ));
-    }));
-
-    // the format! makes sure that the resulting string has exactly 6 characters.
-    format!("{:06}", (rand % 1_000_000))
-}
-
-/// Removes __all__ expired device registrations -> there is no need to check expiration immediately after pruning.
-fn prune_expired_tentative_device_registrations(
-    registrations: &mut HashMap<UserNumber, TentativeDeviceRegistration>,
-) {
-    let now = time();
-
-    registrations.retain(|_, value| match &value {
-        TentativeDeviceRegistration { expiration, .. } => expiration > &now,
-    })
-}
-
-#[update]
-async fn register(device_data: DeviceData, challenge_result: ChallengeAttempt) -> RegisterResponse {
-    if let Err(()) = check_challenge(challenge_result) {
-        return RegisterResponse::BadChallenge;
-    }
-
-    check_device(&device_data, &vec![]);
-
-    if caller() != Principal::self_authenticating(device_data.pubkey.clone()) {
-        trap(&format!(
-            "{} could not be authenticated against {:?}",
-            caller(),
-            device_data.pubkey
-        ));
-    }
-
-    state::ensure_salt_set().await;
-    delegation::prune_expired_signatures();
-
-    let allocation = state::storage_mut(|storage| storage.allocate_user_number());
-    match allocation {
-        Some(user_number) => {
-            write_anchor_data(user_number, vec![DeviceDataInternal::from(device_data)]);
-            RegisterResponse::Registered { user_number }
-        }
-        None => RegisterResponse::CanisterFull,
-    }
-}
-
-#[update]
-async fn add(user_number: UserNumber, device_data: DeviceData) {
-    const MAX_ENTRIES_PER_USER: usize = 10;
-
-    let mut entries = state::anchor_devices(user_number);
-    // must be called before the first await because it requires caller()
-    trap_if_not_authenticated(entries.iter().map(|e| &e.pubkey));
-    state::ensure_salt_set().await;
-
-    check_device(&device_data, &entries);
-
-    if entries
-        .iter()
-        .find(|e| e.pubkey == device_data.pubkey)
-        .is_some()
-    {
-        trap("Device already added.");
-    }
-
-    if entries.len() >= MAX_ENTRIES_PER_USER {
-        trap(&format!(
-            "at most {} authentication information entries are allowed per user",
-            MAX_ENTRIES_PER_USER,
-        ));
-    }
-
-    entries.push(DeviceDataInternal::from(device_data));
-    write_anchor_data(user_number, entries);
-
-    delegation::prune_expired_signatures();
-}
-
-/// Replace or remove an existing device.
-///
-/// NOTE: all mutable operations should call this function because it handles device protection
-fn mutate_device_or_trap(
-    entries: &mut Vec<DeviceDataInternal>,
-    device_key: DeviceKey,
-    new_value: Option<DeviceData>,
-) {
-    let index = match entries.iter().position(|e| e.pubkey == device_key) {
-        None => trap("Could not find device to mutate, check device key"),
-        Some(index) => index,
-    };
-
-    let device = entries.get_mut(index).unwrap();
-
-    // Run appropriate checks for protected devices
-    match device.protection {
-        None => (),
-        Some(DeviceProtection::Unprotected) => (),
-        Some(DeviceProtection::Protected) => {
-            // If the call is not authenticated with the device to mutate, abort
-            if caller() != Principal::self_authenticating(&device.pubkey) {
-                trap("Device is protected. Must be authenticated with this device to mutate");
-            }
-        }
-    };
-
-    match new_value {
-        Some(device_data) => {
-            *device = device_data.into();
-        }
-        None => {
-            // NOTE: we void the more efficient remove_swap to ensure device ordering
-            // is not changed
-            entries.remove(index);
-        }
-    }
-}
-
-#[update]
-async fn update(user_number: UserNumber, device_key: DeviceKey, device_data: DeviceData) {
-    if device_key != device_data.pubkey {
-        trap("device key may not be updated");
-    }
-    let mut entries = state::anchor_devices(user_number);
-
-    trap_if_not_authenticated(entries.iter().map(|e| &e.pubkey));
-    check_device(&device_data, &entries);
-
-    mutate_device_or_trap(&mut entries, device_key, Some(device_data));
-
-    write_anchor_data(user_number, entries);
-
-    delegation::prune_expired_signatures();
-}
-
-#[update]
-async fn remove(user_number: UserNumber, device_key: DeviceKey) {
-    let mut entries = state::anchor_devices(user_number);
-    // must be called before the first await because it requires caller()
-    trap_if_not_authenticated(entries.iter().map(|e| &e.pubkey));
-
-    state::ensure_salt_set().await;
-    delegation::prune_expired_signatures();
-
-    mutate_device_or_trap(&mut entries, device_key, None);
-    write_anchor_data(user_number, entries);
-}
-
-/// Writes the supplied entries to stable memory and updates the anchor operation metric.
-fn write_anchor_data(user_number: UserNumber, entries: Vec<DeviceDataInternal>) {
-    state::storage_mut(|storage| {
-        storage.write(user_number, entries).unwrap_or_else(|err| {
-            trap(&format!(
-                "failed to write device data of user {}: {}",
-                user_number, err
-            ))
-        });
-    });
-
-    state::usage_metrics_mut(|metrics| {
-        metrics.anchor_operation_counter += 1;
-    });
+    tentative_device_registration::verify_tentative_device(user_number, user_verification_code)
+        .await
 }
 
 #[update]
 async fn create_challenge() -> Challenge {
-    let mut rng = make_rng().await;
-
-    delegation::prune_expired_signatures();
-
-    state::inflight_challenges_mut(|inflight_challenges| {
-        let now = time() as u64;
-
-        // Prune old challenges. This drops all challenges that are older than
-        // CAPTCHA_CHALLENGE_LIFETIME
-        inflight_challenges.retain(|_, v| v.created > now - CAPTCHA_CHALLENGE_LIFETIME);
-
-        // Error out if there are too many inflight challenges
-        if inflight_challenges.len() >= MAX_INFLIGHT_CHALLENGES {
-            trap("too many inflight captchas");
-        }
-
-        // actually create the challenge
-
-        // First, we try to find a new (unique) challenge key. It's unlikely we'll have collisions
-        // when generating the key, but to err on the safe side we try up to 10 times.
-        const MAX_TRIES: u8 = 10;
-
-        for _ in 0..MAX_TRIES {
-            let challenge_key = random_string(&mut rng, 10);
-            if !inflight_challenges.contains_key(&challenge_key) {
-                // Then we create the CAPTCHA
-                let (Base64(png_base64), chars) = create_captcha(rng);
-
-                // Finally insert
-                inflight_challenges.insert(
-                    challenge_key.clone(),
-                    ChallengeInfo {
-                        created: now,
-                        chars,
-                    },
-                );
-
-                return Challenge {
-                    png_base64,
-                    challenge_key,
-                };
-            }
-        }
-
-        trap(&format!(
-            "Could not find a new key after {} tries",
-            MAX_TRIES
-        ));
-    })
+    anchor_management::registration::create_challenge().await
 }
 
-// Generate an n-char long string of random characters. The characters are sampled from the rang
-// a-z.
-//
-// NOTE: The 'rand' crate (currently) does not build on wasm32-unknown-unknown so we have to
-// make-do with the RngCore trait (as opposed to Rng), therefore we have to implement this
-// ourselves as opposed to using one of rand's distributions.
-fn random_string<T: RngCore>(rng: &mut T, n: usize) -> String {
-    let mut chars: Vec<u8> = vec![];
-
-    // The range
-    let a: u8 = 'a' as u8;
-    let z: u8 = 'z' as u8;
-
-    // n times, get a random number as u32, then shrink to u8, and finally shrink to the size of
-    // our range. Finally, offset by the start of our range.
-    for _ in 0..n {
-        let next: u8 = rng.next_u32() as u8 % (z - a) + a;
-        chars.push(next);
-    }
-
-    return String::from_utf8_lossy(&chars).to_string();
+#[update]
+async fn register(device_data: DeviceData, challenge_result: ChallengeAttempt) -> RegisterResponse {
+    anchor_management::registration::register(device_data, challenge_result).await
 }
 
-// Get a random number generator based on 'raw_rand'
-async fn make_rng() -> rand_chacha::ChaCha20Rng {
-    let raw_rand: Vec<u8> = match call(Principal::management_canister(), "raw_rand", ()).await {
-        Ok((res,)) => res,
-        Err((_, err)) => trap(&format!("failed to get seed: {}", err)),
-    };
-    let seed: Salt = raw_rand[..].try_into().unwrap_or_else(|_| {
-        trap(&format!(
-                "when creating seed from raw_rand output, expected raw randomness to be of length 32, got {}",
-                raw_rand.len()
-                ));
-    });
-
-    rand_chacha::ChaCha20Rng::from_seed(seed)
+#[update]
+async fn add(user_number: UserNumber, device_data: DeviceData) {
+    anchor_management::add(user_number, device_data).await
 }
 
-#[cfg(feature = "dummy_captcha")]
-fn create_captcha<T: RngCore>(rng: T) -> (Base64, String) {
-    let mut captcha = captcha::RngCaptcha::from_rng(rng);
-    let captcha = captcha.set_chars(&vec!['a']).add_chars(1).view(96, 48);
-
-    let resp = match captcha.as_base64() {
-        Some(png_base64) => Base64(png_base64),
-        None => trap("Could not get base64 of captcha"),
-    };
-
-    return (resp, captcha.chars_as_string());
+#[update]
+async fn update(user_number: UserNumber, device_key: DeviceKey, device_data: DeviceData) {
+    anchor_management::update(user_number, device_key, device_data).await
 }
 
-#[cfg(not(feature = "dummy_captcha"))]
-fn create_captcha<T: RngCore>(rng: T) -> (Base64, String) {
-    let mut captcha = captcha::RngCaptcha::from_rng(rng);
-    let captcha = captcha
-        .add_chars(5)
-        .apply_filter(Wave::new(2.0, 20.0).horizontal())
-        .apply_filter(Wave::new(2.0, 20.0).vertical())
-        .view(220, 120);
-
-    let resp = match captcha.as_base64() {
-        Some(png_base64) => Base64(png_base64),
-        None => trap("Could not get base64 of captcha"),
-    };
-
-    return (resp, captcha.chars_as_string());
-}
-
-// Check whether the CAPTCHA challenge was solved
-fn check_challenge(res: ChallengeAttempt) -> Result<(), ()> {
-    state::inflight_challenges_mut(|inflight_challenges| {
-        match inflight_challenges.remove(&res.key) {
-            Some(challenge) => {
-                if res.chars != challenge.chars {
-                    return Err(());
-                }
-                return Ok(());
-            }
-            None => Err(()),
-        }
-    })
+#[update]
+async fn remove(user_number: UserNumber, device_key: DeviceKey) {
+    anchor_management::remove(user_number, device_key).await
 }
 
 /// Returns all devices of the user (authentication and recovery) but no information about device registrations.
@@ -514,42 +98,7 @@ fn lookup(user_number: UserNumber) -> Vec<DeviceData> {
 
 #[update] // this is an update call because queries are not (yet) certified
 fn get_anchor_info(user_number: UserNumber) -> IdentityAnchorInfo {
-    let entries = state::anchor_devices(user_number);
-    trap_if_not_authenticated(entries.iter().map(|e| &e.pubkey));
-
-    let devices = entries.into_iter().map(DeviceData::from).collect();
-    let now = time();
-
-    state::tentative_device_registrations(|tentative_device_registrations| {
-        match tentative_device_registrations.get(&user_number) {
-            Some(TentativeDeviceRegistration {
-                expiration,
-                state:
-                    DeviceTentativelyAdded {
-                        tentative_device, ..
-                    },
-            }) if *expiration > now => IdentityAnchorInfo {
-                devices,
-                device_registration: Some(DeviceRegistrationInfo {
-                    expiration: *expiration,
-                    tentative_device: Some(tentative_device.clone()),
-                }),
-            },
-            Some(TentativeDeviceRegistration { expiration, .. }) if *expiration > now => {
-                IdentityAnchorInfo {
-                    devices,
-                    device_registration: Some(DeviceRegistrationInfo {
-                        expiration: *expiration,
-                        tentative_device: None,
-                    }),
-                }
-            }
-            None | Some(_) => IdentityAnchorInfo {
-                devices,
-                device_registration: None,
-            },
-        }
-    })
+    anchor_management::get_anchor_info(user_number)
 }
 
 #[query]
@@ -642,72 +191,6 @@ fn trap_if_not_authenticated<'a>(public_keys: impl Iterator<Item = &'a PublicKey
         }
     }
     trap(&format!("{} could not be authenticated.", caller()))
-}
-
-/// This checks some device invariants, in particular:
-///   * Sizes of various fields do not exceed limits
-///   * Only recovery phrases can be protected
-///   * There can only be one recovery phrase
-///
-///  Otherwise, trap.
-///
-///  NOTE: while in the future we may lift this restriction, for now we do ensure that
-///  protected devices are limited to recovery phrases, which the webapp expects.
-fn check_device(device_data: &DeviceData, existing_devices: &[DeviceDataInternal]) {
-    check_entry_limits(device_data);
-
-    if device_data.protection == DeviceProtection::Protected
-        && device_data.key_type != KeyType::SeedPhrase
-    {
-        trap(&format!(
-            "Only recovery phrases can be protected but key type is {:?}",
-            device_data.key_type
-        ));
-    }
-
-    // if the device is a recovery phrase, check if a different recovery phrase already exists
-    if device_data.key_type == KeyType::SeedPhrase
-        && existing_devices.iter().any(|existing_device| {
-            existing_device.pubkey != device_data.pubkey
-                && existing_device.key_type == Some(KeyType::SeedPhrase)
-        })
-    {
-        trap("There is already a recovery phrase and only one is allowed.");
-    }
-}
-
-fn check_entry_limits(device_data: &DeviceData) {
-    const ALIAS_LEN_LIMIT: usize = 64;
-    const PK_LEN_LIMIT: usize = 300;
-    const CREDENTIAL_ID_LEN_LIMIT: usize = 200;
-
-    let n = device_data.alias.len();
-    if n > ALIAS_LEN_LIMIT {
-        trap(&format!(
-            "alias length {} exceeds the limit of {} bytes",
-            n, ALIAS_LEN_LIMIT,
-        ));
-    }
-
-    let n = device_data.pubkey.len();
-    if n > PK_LEN_LIMIT {
-        trap(&format!(
-            "public key length {} exceeds the limit of {} bytes",
-            n, PK_LEN_LIMIT,
-        ));
-    }
-
-    let n = device_data
-        .credential_id
-        .as_ref()
-        .map(|bytes| bytes.len())
-        .unwrap_or_default();
-    if n > CREDENTIAL_ID_LEN_LIMIT {
-        trap(&format!(
-            "credential id length {} exceeds the limit of {} bytes",
-            n, CREDENTIAL_ID_LEN_LIMIT,
-        ));
-    }
 }
 
 fn main() {}
